IEXHaskellSDK.Core

Description

IEXHaskellSDKConfig

data IEXHaskellSDKConfig

Constructors

IEXHaskellSDKConfig

Fields

configHost :: ByteString

host supplied in the Request

configUserAgent :: Text

user-agent supplied in the Request

configLogExecWithContext :: LogExecWithContext

Run a block using a Logger instance

configLogContext :: LogContext

Configures the logger

configAuthMethods :: [AnyAuthMethod]

List of configured auth methods

configValidateAuthMethods :: Bool

throw exceptions if auth methods are not configured

newConfig :: IO IEXHaskellSDKConfig

constructs a default IEXHaskellSDKConfig

configHost:

http://api.iextrading.com/1.0

configUserAgent:

"iex-haskell-sdk/0.1.0.0"

addAuthMethod :: AuthMethod auth => IEXHaskellSDKConfig -> auth -> IEXHaskellSDKConfig

updates config use AuthMethod on matching requests

withStdoutLogging :: IEXHaskellSDKConfig -> IO IEXHaskellSDKConfig

updates the config to use stdout logging

withStderrLogging :: IEXHaskellSDKConfig -> IO IEXHaskellSDKConfig

updates the config to use stderr logging

withNoLogging :: IEXHaskellSDKConfig -> IEXHaskellSDKConfig

updates the config to disable logging

IEXHaskellSDKRequest

data IEXHaskellSDKRequest req contentType res accept

Represents a request.

Type Variables:

req - request operation

contentType - MimeType associated with request body

res - response model

accept - MimeType associated with response body

Constructors

IEXHaskellSDKRequest

Fields

rMethod :: Method

Method of IEXHaskellSDKRequest

rUrlPath :: [ByteString]

Endpoint of IEXHaskellSDKRequest

rParams :: Params

params of IEXHaskellSDKRequest

rAuthTypes :: [TypeRep]

types of auth methods

rMethodL :: Lens_' (IEXHaskellSDKRequest req contentType res accept) Method

rMethod Lens

rUrlPathL :: Lens_' (IEXHaskellSDKRequest req contentType res accept) [ByteString]

rUrlPath Lens

rParamsL :: Lens_' (IEXHaskellSDKRequest req contentType res accept) Params

rParams Lens

rAuthTypesL :: Lens_' (IEXHaskellSDKRequest req contentType res accept) [TypeRep]

rParams Lens

HasBodyParam

class HasBodyParam req param where

Designates the body parameter of a request

Minimal complete definition

Nothing

Methods

setBodyParam :: forall contentType res accept. (Consumes req contentType, MimeRender contentType param) => IEXHaskellSDKRequest req contentType res accept -> param -> IEXHaskellSDKRequest req contentType res accept

HasOptionalParam

class HasOptionalParam req param where

Designates the optional parameters of a request

Minimal complete definition

applyOptionalParam | (-&-)

Methods

applyOptionalParam :: IEXHaskellSDKRequest req contentType res accept -> param -> IEXHaskellSDKRequest req contentType res accept

Apply an optional parameter to a request

(-&-) :: IEXHaskellSDKRequest req contentType res accept -> param -> IEXHaskellSDKRequest req contentType res accept infixl 2

infix operator / alias for addOptionalParam

data Params

Request Params

Constructors

Params

Fields

paramsQuery :: Query

paramsHeaders :: RequestHeaders

paramsBody :: ParamBody

paramsQueryL :: Lens_' Params Query

paramsQuery Lens

paramsHeadersL :: Lens_' Params RequestHeaders

paramsHeaders Lens

paramsBodyL :: Lens_' Params ParamBody

paramsBody Lens

data ParamBody

Request Body

Constructors

ParamBodyNone

ParamBodyB ByteString

ParamBodyBL ByteString

ParamBodyFormUrlEncoded Form

ParamBodyMultipartFormData [Part]

IEXHaskellSDKRequest Utils

_mkRequest

Arguments

:: Method
Method

-> [ByteString]
Endpoint

-> IEXHaskellSDKRequest req contentType res accept
req: Request Type, res: Response Type

_mkParams :: Params

setHeader :: IEXHaskellSDKRequest req contentType res accept -> [Header] -> IEXHaskellSDKRequest req contentType res accept

removeHeader :: IEXHaskellSDKRequest req contentType res accept -> [HeaderName] -> IEXHaskellSDKRequest req contentType res accept

_setContentTypeHeader :: forall req contentType res accept. MimeType contentType => IEXHaskellSDKRequest req contentType res accept -> IEXHaskellSDKRequest req contentType res accept

_setAcceptHeader :: forall req contentType res accept. MimeType accept => IEXHaskellSDKRequest req contentType res accept -> IEXHaskellSDKRequest req contentType res accept

setQuery :: IEXHaskellSDKRequest req contentType res accept -> [QueryItem] -> IEXHaskellSDKRequest req contentType res accept

addForm :: IEXHaskellSDKRequest req contentType res accept -> Form -> IEXHaskellSDKRequest req contentType res accept

_addMultiFormPart :: IEXHaskellSDKRequest req contentType res accept -> Part -> IEXHaskellSDKRequest req contentType res accept

_setBodyBS :: IEXHaskellSDKRequest req contentType res accept -> ByteString -> IEXHaskellSDKRequest req contentType res accept

_setBodyLBS :: IEXHaskellSDKRequest req contentType res accept -> ByteString -> IEXHaskellSDKRequest req contentType res accept

_hasAuthType :: AuthMethod authMethod => IEXHaskellSDKRequest req contentType res accept -> Proxy authMethod -> IEXHaskellSDKRequest req contentType res accept

Params Utils

toPath :: ToHttpApiData a => a -> ByteString

toHeader :: ToHttpApiData a => (HeaderName, a) -> [Header]

toForm :: ToHttpApiData v => (ByteString, v) -> Form

toQuery :: ToHttpApiData a => (ByteString, Maybe a) -> [QueryItem]

OpenAPI CollectionFormat Utils

data CollectionFormat

Determines the format of the array if type array is used.

Constructors

CommaSeparated
CSV format for multiple parameters.

SpaceSeparated
Also called SSV

TabSeparated
Also called TSV

PipeSeparated
`value1|value2|value2`

MultiParamArray
Using multiple GET parameters, e.g. `foo=bar&foo=baz`. This is valid only for parameters in "query" (Query) or "formData" (Form)

toHeaderColl :: ToHttpApiData a => CollectionFormat -> (HeaderName, [a]) -> [Header]

toFormColl :: ToHttpApiData v => CollectionFormat -> (ByteString, [v]) -> Form

toQueryColl :: ToHttpApiData a => CollectionFormat -> (ByteString, Maybe [a]) -> Query

_toColl :: Traversable f => CollectionFormat -> (f a -> [(b, ByteString)]) -> f [a] -> [(b, ByteString)]

_toCollA :: (Traversable f, Traversable t, Alternative t) => CollectionFormat -> (f (t a) -> [(b, t ByteString)]) -> f (t [a]) -> [(b, t ByteString)]

_toCollA' :: (Monoid c, Traversable f, Traversable t, Alternative t) => CollectionFormat -> (f (t a) -> [(b, t c)]) -> (Char -> c) -> f (t [a]) -> [(b, t c)]

AuthMethods

class Typeable a => AuthMethod a where

Provides a method to apply auth methods to requests

Methods

applyAuthMethod :: IEXHaskellSDKConfig -> a -> IEXHaskellSDKRequest req contentType res accept -> IO (IEXHaskellSDKRequest req contentType res accept)

data AnyAuthMethod

An existential wrapper for any AuthMethod

Constructors

forall a.AuthMethod a => AnyAuthMethod a

data AuthMethodException

indicates exceptions related to AuthMethods

Constructors

AuthMethodException String

_applyAuthMethods :: IEXHaskellSDKRequest req contentType res accept -> IEXHaskellSDKConfig -> IO (IEXHaskellSDKRequest req contentType res accept)

apply all matching AuthMethods in config to request

Utils

_omitNulls :: [(Text, Value)] -> Value

Removes Null fields. (OpenAPI-Specification 2.0 does not allow Null in JSON)

_toFormItem :: (ToHttpApiData a, Functor f) => t -> f a -> f (t, [Text])

Encodes fields using WH.toQueryParam

_emptyToNothing :: Maybe String -> Maybe String

Collapse (Just "") to Nothing

_memptyToNothing :: (Monoid a, Eq a) => Maybe a -> Maybe a

Collapse (Just mempty) to Nothing

DateTime Formatting

newtype DateTime

Constructors

DateTime

Fields

unDateTime :: UTCTime

_readDateTime :: (MonadFail m, Alternative m) => String -> m DateTime

_parseISO8601

_showDateTime :: (t ~ UTCTime, FormatTime t) => t -> String

TI.formatISO8601Millis

_parseISO8601 :: (ParseTime t, MonadFail m, Alternative m) => String -> m t

parse an ISO8601 date-time string

Date Formatting

newtype Date

Constructors

Date

Fields

unDate :: Day

_readDate :: MonadFail m => String -> m Date

TI.parseTimeM True TI.defaultTimeLocale "%Y-%m-%d"

_showDate :: FormatTime t => t -> String

TI.formatTime TI.defaultTimeLocale "%Y-%m-%d"

Byte/Binary Formatting

newtype ByteArray

base64 encoded characters

Constructors

ByteArray

Fields

unByteArray :: ByteString

_readByteArray :: MonadFail m => Text -> m ByteArray

read base64 encoded characters

_showByteArray :: ByteArray -> Text

show base64 encoded characters

newtype Binary

any sequence of octets

Constructors

Binary

Fields

unBinary :: ByteString

_readBinaryBase64 :: MonadFail m => Text -> m Binary

_showBinaryBase64 :: Binary -> Text

Lens Type Aliases

type Lens_' s a = Lens_ s s a a

type Lens_ s t a b = forall (f :: * -> *). Functor f => (a -> f b) -> s -> f t